rename collectionTable.tsx collectionsTable.tsx
